requirements.txt
SpeechRecognition
moviepy==1.0.3
pydub
pydub
openai-whisper